Implement SipHash PRF
    This commit implements all SipHash available primitives.
The kryptos_siphash_sum() primitive can be useful for people seeking to
use SipHash as a hash function for hash table. The function macro
kryptos_run_cipher_siphash() is useful for people that want to use
SipHash as a MAC algorithm.

    Although its name, SipHash is NOT a cryptographic hash function
due to it some care for not allowing its use as a crypto hash was taken.
It does not implement hash_size() function.

    All required unit tests were implemented (user, kernel).

    Speed up build option --no-siphash-tests implemented.

    New samples were added to the sample base.

    All documentation was updated.

* Fix: Wrong calls to triple_des inside triple_des_ede integration tests
with Poly1305 were corrected accordingly (user, kernel).