Implement new CI based on github actions
This new CI includes:

- Linux GCC/Clang
- FreeBSD Clang/GCC
- NetBSD GCC/Clang
- OpenBSD Clang/GCC
- Windows GCC/MSVC

Clang on Windows is not my goal. My main concern on Clang is to provide
a functional build for platforms that use it as the default C compiler
(such as FreeBSD) and where the development is really mature. Clang on
Windows (from VS) is buggy, specially when dealing with int scalar type
extension. It is broken since 2015!!! It compiles but not links out.
Nevermind, crappy.